Refactored significance calculation into separate file
